reactor: provider 重构
