+++ web/src/modules/layout/AdaptiveMenuLayout.tsx
@@ -1,5 +1,6 @@
-import React, { ReactNode, useEffect, useState } from "react";
+import React, { ReactNode } from "react";
 import styled from "@emotion/styled";
+import { IS_MOBILE_LAYOUT, useLayoutDirection, LayoutDirection } from "@/global-provider.tsx";
 
 export type AdaptiveMenuLayoutProps = {
   menu?: ReactNode | ((direction: LayoutDirection) => ReactNode);
@@ -7,43 +8,21 @@ export type AdaptiveMenuLayoutProps = {
   style?: React.CSSProperties;
   className?: string;
 };
-export const IS_MOBILE_LAYOUT = "(max-width: 550px) and (orientation: portrait)";
-function getDefaultDirection() {
-  return globalThis.matchMedia?.(IS_MOBILE_LAYOUT).matches ? LayoutDirection.Vertical : LayoutDirection.Horizontal;
-}
+
 export function AdaptiveMenuLayout(props: AdaptiveMenuLayoutProps) {
   const { className, style } = props;
   let { menu, children } = props;
 
-  const [direction, setDirection] = useState(getDefaultDirection);
-  useEffect(() => {
-    const mql = window.matchMedia(IS_MOBILE_LAYOUT);
-    const direction = mql.matches ? LayoutDirection.Vertical : LayoutDirection.Horizontal;
-    if (mql.matches) setDirection(direction);
-    const onChange = (e: MediaQueryListEvent) => {
-      setDirection(e.matches ? LayoutDirection.Vertical : LayoutDirection.Horizontal);
-    };
-    mql.addEventListener("change", onChange);
-    return () => {
-      mql.removeEventListener("change", onChange);
-    };
-  }, []);
+  const direction = useLayoutDirection();
   if (typeof menu === "function") menu = menu(direction);
   if (typeof children === "function") children = children(direction);
   return (
-    <AdaptiveLayoutContext value={direction}>
-      <AdaptiveMenuLayoutCSS className={className} style={style}>
-        <div className="adaptive-menu">{menu}</div>
-        <div className="adaptive-content">{children}</div>
-      </AdaptiveMenuLayoutCSS>
-    </AdaptiveLayoutContext>
+    <AdaptiveMenuLayoutCSS className={className} style={style}>
+      <div className="adaptive-menu">{menu}</div>
+      <div className="adaptive-content">{children}</div>
+    </AdaptiveMenuLayoutCSS>
   );
 }
-export enum LayoutDirection {
-  Horizontal = 0,
-  Vertical = 1,
-}
-export const AdaptiveLayoutContext = React.createContext<LayoutDirection>(LayoutDirection.Horizontal);
 
 const AdaptiveMenuLayoutCSS = styled.div`
   display: flex;
