+++ web/src/modules/post/components/PostContent.tsx
@@ -1,5 +1,5 @@
 import { AssetImage, AssetMediaDto, AssetVideo, TextStructure } from "@/api.ts";
-import { useThemeToken } from "@/hooks/antd.ts";
+import { useThemeToken } from "@/global-provider.tsx";
 import styled from "@emotion/styled";
 import React, { CSSProperties, useMemo } from "react";
 import { ReactNode } from "react";
