+++ web/src/modules/post/components/PostHeader.tsx
@@ -1,4 +1,4 @@
-import { useThemeToken } from "@/hooks/antd.ts";
+import { useThemeToken } from "@/global-provider.tsx";
 import styled from "@emotion/styled";
 import { Space } from "antd";
 import React, { CSSProperties } from "react";
